+++ src/test/polymer/polymer2-element-scanner_test.ts
@@ -488,6 +488,23 @@ namespaced name.`,
           ],
           observerProperties: [['foo'], ['foo', 'bar']],
         }]);
+  });
+
+  test('can identify elements registered with ClassName.is', async() => {
+    const elements = await getElements('test-element-11.js');
+    const elementData = await Promise.all(elements.map(getTestProps));
+    assert.deepEqual(
+        elementData, [{
+          attributes: [{name: 'prop1'}],
+          className: 'MyElement',
+          description: '',
+          methods: [],
+          properties: [{name: 'prop1', description: '', type: 'string'}],
+          summary: '',
+          superClass: 'Polymer.Element',
+          tagName: 'my-app',
+          warningUnderlines: [],
+        }]);
 
   });
 });